@@ -1,6 +1,6 @@
 {
   "name": "react-data-table-component",
-  "version": "0.8.0",
+  "version": "0.9.0",
   "description": "A declarative react based data table",
   "main": "dist/index.prod.js",
   "module": "dist/index.prod.js",
@@ -90,8 +90,8 @@
     "jest": "^22.4.3",
     "jest-styled-components": "^5.0.0",
     "raf": "^3.4.0",
-    "react": "^16.3.0",
-    "react-dom": "^16.3.0",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
     "rimraf": "^2.6.2",
     "rollup": "^0.57.1",
     "rollup-plugin-babel": "^3.0.3",
@@ -104,10 +104,12 @@
     "stylelint-config-standard": "^18.2.0",
     "stylelint-config-styled-components": "^0.1.1",
     "stylelint-processor-styled-components": "^1.3.1",
-    "uglify-es": "^3.3.9"
+    "uglify-es": "^3.3.9",
+    "yarn": "^1.7.0"
   },
   "dependencies": {
     "lodash": "^4.17.5",
+    "react-lifecycles-compat": "^3.0.4",
     "shortid": "^2.2.8"
   },
   "peerDependencies": {
