@@ -74,8 +74,9 @@ export const handleSort = ({ expandableRows, expanderStateField }, selector, sor
   return { ...state };
 };
 
-export const clearSelectedRows = () => ({
+export const clearSelectedRows = clearedRowsFlag => ({
   allSelected: false,
   selectedCount: 0,
   selectedRows: [],
+  clearSelectedRows: clearedRowsFlag,
 });
